+++ dev/user.clj
@@ -7,30 +7,19 @@
 ;;------
 (comment
   (reload/init
-   {:dirs ["src" "dev" "test"]})
+   {:dirs ["src" "dev" "test"]
+    :no-reload '#{user}})
 
   (reload/reload)
   :rcf)
 ;;------
-
-(comment
-
-  (defn f
-    ([] (f {}))
-    ([{:keys [schema storage]}]
-     [schema storage]))
-
-  (f {:schema 1 :storage 2})
-  ;; => [1 2]
-  (f)
-  ;; => [nil nil]
-  :rcf)
-
-;;------
-
 (comment
 
   (start)
+
+  (def now (jt/instant))
+
+  (gc)
 
   (ds/puts! [{:db/id 250, :name "250", :friends 251}
              {:db/id 251, :name "251"}])
@@ -41,7 +30,9 @@
           [?e :friends ?f]])
 
   (get (ds/entity 11) :friends)
+
   (stop)
+
   (conn?)
 
   :rcf)
@@ -50,14 +41,33 @@
 (comment
   (def schema {:aka {:db/cardinality :db.cardinality/many}})
 
-  (start schema)
+  (start {:schema schema})
 
   (ds/puts! [{:db/id -1
               :name  "Maksim"
               :age   45
               :aka   ["Max Otto von Stierlitz", "Jack Ryan"]}])
+
   (ds/q '[:find  ?n ?a
           :where [?e :aka "Max Otto von Stierlitz"]
           [?e :name ?n]
           [?e :age  ?a]])
+
+  (stop)
+
+  :rcf)
+;;-----
+(comment
+  (start {:url "jdbc:sqlite:storage/db.sqlite"})
+  (ds/puts! [{:db/id -1, :now (jt/instant)}])
+  (-> (ds/q '[:find ?e ?time
+              :keys e   time
+              :where
+              [?e :now ?time]])
+      first
+      :time
+      str)
+
+  (stop)
+  (conn?)
   :rcf)