+++ src/hkimjp/datascript.clj
@@ -11,7 +11,7 @@
 (time-literals.read-write/print-time-literals-clj!)
 
 (def conn nil)
-(def ^:private storage nil)
+(def storage nil)
 
 (defn- datasource
   ([] (datasource "jdbc:sqlite:data/db.sqlite"))
@@ -60,13 +60,21 @@
 
 ;; ------------------------------
 
+;; (defn start
+;;   ([] (create-conn nil nil))
+;;   ([schema] (create-conn schema nil))
+;;   ([schema url]
+;;    (if (exist? url)
+;;      (restore-conn (make-storage url))
+;;      (create-conn schema {:storage (make-storage url)}))))
+
 (defn start
   ([] (create-conn nil nil))
-  ([schema] (create-conn schema nil))
-  ([schema url]
-   (if (exist? url)
-     (restore-conn (make-storage url))
-     (create-conn schema {:storage (make-storage url)}))))
+  ([{:keys [schema url]}]
+   (cond
+     (nil? url) (create-conn schema)
+     (exist? url) (restore-conn (make-storage url))
+     :else (create-conn schema {:storage (make-storage url)}))))
 
 (defn stop []
   (close-conn))
@@ -75,27 +83,25 @@
   (d/conn? conn))
 
 (defn gc []
-  (d/collect-garbage storage))
+  (when (some? storage)
+    (d/collect-garbage storage)))
 
 ;;-----------------------------
 
 (defn- abbrev
+  "shorten string for concise log."
   ([s] (abbrev s 80))
   ([s n] (let [pat (re-pattern (str "(^.{" n "}).*"))]
            (str/replace-first s pat "$1..."))))
 
-;; this did not work with private conn.
+;; this did not work with (def ^:private conn nil)
 ;; (defmacro q [query & inputs]
 ;;   (t/log! :info (str "q " query))
 ;;   `(d/q ~query @conn ~@inputs))
 
 (defn q [query & inputs]
   (t/log! :info (str "q " query))
-  `(d/q ~query @conn ~@inputs))
-
-;; (defn q [query & inputs]
-;;   (t/log! :info (str "q " query))
-;;   (apply d/q query @conn inputs))
+  (apply d/q query @conn inputs))
 
 (defn entity [id]
   (d/entity @conn id))